+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/session/DefaultMeetingSession.kt
@@ -76,7 +76,8 @@ class DefaultMeetingSession(
                 logger,
                 turnRequestParams,
                 metricsCollector,
-                videoClientStateController
+                videoClientStateController,
+                configuration.urls.urlRewriter
             )
 
         val videoClientController =